requirements.txt
paddlepaddle
paddlehub
visualdl
paddleseg
paddlenlp
paddleocr
opencv-python
mmcv-full